refactor(counting-section.astro): rebuild counting form with correct components
