+++ src/components/organisms/sections/counting-section.astro
@@ -21,12 +21,12 @@ import { t } from '@utils/lang-utils'
 
         <!-- Arrived -->
         <div class="sm:col-span-4">
-          <Datepicker name={t('home:counting.arrived')} required />
+          <Datepicker name={t('home:counting.arrival')} required />
         </div>
 
         <!-- Departed -->
         <div class="sm:col-span-4">
-          <Datepicker name={t('home:counting.departed')} required />
+          <Datepicker name={t('home:counting.departure')} required />
         </div>
       </div>
     </div>
