+++ src/pages/index.astro
@@ -2,7 +2,8 @@
 import { changeLanguage, t } from 'i18next'
 
 import Hero from '@atoms/hero.astro'
-import CountingSection from '@atoms/sections/counting-section.astro'
+
+import CountingSection from '@organisms/sections/counting-section.astro'
 
 import HtmlLayout from '@layouts/html-layout.astro'
 import MainLayout from '@layouts/main-layout.astro'
